[Added] Get as type to INISection and XMLElement [Added] Settings framework [Changed] Time step to use f64 (or double) to avoid precision loss [Changed] has checks in XMLElement to avoid unsafe bool conversions [Changed] Cleaned up code

+++ Source/Eris/Application/Events.h
@@ -36,7 +36,7 @@ namespace Eris
         EVENT(UpdateEvent)
 
     public:
-        glm::f32 time_step;
+        glm::f64 time_step;
     };
 
     struct PostUpdateEvent : public Event
